Moved token refresh and validation to middleware
The token validation and refresh are both performed in the same middleware function.

If the token expired, send back an error that will be used to logout the person on the frontend.
If the token is valid, refresh the token, update the "user" value so it's up to date with the database after every action.